# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ProjectTnW - Multi-tenant School Transportation SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user and create a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user and their tenants
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/switch-tenant:
    post:
      operationId: switchTenant
      tags: [auth]
      summary: Switch active tenant context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchTenantBody"
      responses:
        "200":
          description: Switched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── TENANTS ───
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List drivers
      responses:
        "200":
          description: Drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"

    post:
      operationId: createDriver
      tags: [drivers]
      summary: Create a driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDriverBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{driverId}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get a driver
      parameters:
        - in: path
          name: driverId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateDriver
      tags: [drivers]
      summary: Update a driver
      parameters:
        - in: path
          name: driverId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

    delete:
      operationId: deleteDriver
      tags: [drivers]
      summary: Delete a driver
      parameters:
        - in: path
          name: driverId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── ROUTES ───
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        tenantName:
          type: string
        tenantSlug:
          type: string
      required:
        - email
        - password
        - fullName
        - tenantName
        - tenantSlug
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    SwitchTenantBody:
      type: object
      properties:
        tenantId:
          type: string
      required:
        - tenantId
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
        tenant:
          $ref: "#/components/schemas/Tenant"
        role:
          type: string
      required:
        - token
        - user
        - tenant
        - role
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]
        isActive:
          type: boolean
        isSuperAdmin:
          type: boolean
        createdAt:
          type: ["string", "null"]
      required:
        - id
        - email
        - isActive
        - isSuperAdmin
    MeResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        tenants:
          type: array
          items:
            $ref: "#/components/schemas/UserTenantMembership"
        currentTenant:
          $ref: "#/components/schemas/Tenant"
        currentRole:
          type: string
      required:
        - user
        - tenants
    UserTenantMembership:
      type: object
      properties:
        tenantId:
          type: string
        tenantName:
          type: string
        tenantSlug:
          type: string
        role:
          type: string
      required:
        - tenantId
        - tenantName
        - tenantSlug
        - role

    # ─── TENANT ───
    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        isDemo:
          type: ["boolean", "null"]
        plan:
          type: ["string", "null"]
        billingStatus:
          type: ["string", "null"]
        trialEndsAt:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
      required:
        - id
        - name
        - slug
...
    Driver:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        userId:
          type: ["string", "null"]
        licenseNumber:
          type: string
        phone:
          type: ["string", "null"]
        hireDate:
          type: ["string", "null"]
        assignedVehicleId:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        vehicleLicensePlate:
          type: ["string", "null"]
      required:
        - id
        - tenantId
        - licenseNumber
    CreateDriverBody:
      type: object
      properties:
        userId:
          type: string
        licenseNumber:
          type: string
        phone:
          type: string
        hireDate:
          type: string
        assignedVehicleId:
          type: string
      required:
        - licenseNumber
    UpdateDriverBody:
      type: object
      properties:
        licenseNumber:
          type: string
        phone:
          type: string
        hireDate:
          type: string
        assignedVehicleId:
          type: ["string", "null"]

    # ─── ROUTES ───
...